// src/domain/models/PersonModel.ts
export type PersonModel = {
  id?: string
  name: string
  email: string
  isOwner: boolean
  document: string
  kin: string
  occupation: string
  wage: number
  wageSources: string
  nis: string
  schooling: string
  birthdate: Date
  phone: string
  isWhatsApp: boolean
  phone2: string
  isWhatsApp2: boolean
}